requirements.txt
alembic==1.10.2
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.0
python-dotenv
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
alembic==1.10.2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pytest==7.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
